dev_test/shopping_chat_bot/shopping_chatbot/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
requests==2.31.0
python-dotenv==1.0.1

# Use psycopg2-binary if switching to PostgreSQL:
# psycopg2-binary==2.9.9
